requirements.txt
ultralytics>=8.4
opencv-python>=4.8
numpy>=1.24
torch>=2.0
pyyaml>=6.0
tqdm>=4.60
